LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/border-image-outset-expected.txt


PASS Can set 'border-image-outset' to CSS-wide keywords: initial
PASS Can set 'border-image-outset' to CSS-wide keywords: inherit
PASS Can set 'border-image-outset' to CSS-wide keywords: unset
PASS Can set 'border-image-outset' to CSS-wide keywords: revert
FAIL Can set 'border-image-outset' to var() references:  var(--A) assert_equals: expected 2 but got 1
FAIL Can set 'border-image-outset' to a length: 0px assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-outset' to a length: -3.14em assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-outset' to a length: 3.14cm assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-outset' to a length: calc(0px + 0em) assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSStyleValue]"
FAIL Can set 'border-image-outset' to a number: 0 assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-outset' to a number: -3.14 assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-outset' to a number: 3.14 assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-outset' to a number: calc(2 + 3) assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSStyleValue]"
PASS Setting 'border-image-outset' to a percent throws TypeError
PASS Setting 'border-image-outset' to a time throws TypeError
PASS Setting 'border-image-outset' to an angle throws TypeError
PASS Setting 'border-image-outset' to a flexible length throws TypeError
PASS Setting 'border-image-outset' to a URL throws TypeError
PASS Setting 'border-image-outset' to a transform throws TypeError
PASS 'border-image-outset' does not supported '1 1.2'
PASS 'border-image-outset' does not supported '30px 2 45px'
PASS 'border-image-outset' does not supported '7px 12px 14px 5px'
